+++ src/com/company/tax/TaxService.java
@@ -1,7 +1,9 @@
 package com.company.tax;
 
+import java.math.BigDecimal;
+
 public class TaxService {
-    public void payOut(double taxAmount) {
+    public void payOut(BigDecimal taxAmount) {
         System.out.format("Уплачен налог в размере %.2f%n", taxAmount);
     }
 }
